+++ R/descr.R
@@ -10,6 +10,12 @@
 #'    in the output. If variable names are longer than \code{max.length},
 #'    they will be shortened to the last whole word within the first
 #'    \code{max.length} chars.
+#' @param show Character vector, indicating which information (columns) that describe
+#'   the data should be returned. May be one or more of \code{"type", "label", "n",
+#'   "NA.prc", "mean", "sd", "se", "md", "trimmed", "range", "skew"}. There are
+#'   two shortcuts: \code{show = "all"} (default) shows all information,
+#'   \code{show = "short"} just shows n, missing percentage, mean and standard
+#'   deviation.
 #' @param out Character vector, indicating whether the results should be printed
 #'    to console (\code{out = "txt"}) or as HTML-table in the viewer-pane
 #'    (\code{out = "viewer"}) or browser (\code{out = "browser"}).
@@ -33,6 +39,9 @@
 #' library(dplyr)
 #' efc %>% select(e42dep, e15relat, c172code) %>% descr()
 #'
+#' # show just a few elements
+#' efc %>% select(e42dep, e15relat, c172code) %>% descr(show = "short")
+#'
 #' # with grouped data frames
 #' efc %>%
 #'   group_by(e16sex) %>%
@@ -50,7 +59,14 @@
 #' @importFrom dplyr select mutate
 #' @importFrom sjlabelled copy_labels
 #' @export
-descr <- function(x, ..., max.length = NULL, weights = NULL, out = c("txt", "viewer", "browser")) {
+descr <- function(x,
+                  ...,
+                  max.length = NULL,
+                  weights = NULL,
+                  show = "all",
+                  out = c("txt", "viewer", "browser"),
+                  encoding = "UTF-8",
+                  file = NULL) {
 
   out <- match.arg(out)
 
@@ -58,6 +74,15 @@ descr <- function(x, ..., max.length = NULL, weights = NULL, out = c("txt", "vie
     message("Package `sjPlot` needs to be loaded to print HTML tables.")
     out <- "txt"
   }
+
+  # select elements that should be shown
+
+  if (show == "all")
+    show <- c("type", "label", "n", "NA.prc", "mean", "sd", "se", "md", "trimmed", "range", "skew")
+  else if (show == "short")
+    show <- c("n", "NA.prc", "mean", "sd")
+
+  show <- c("var", show)
 
 
   # get dot data
@@ -87,7 +112,7 @@ descr <- function(x, ..., max.length = NULL, weights = NULL, out = c("txt", "vie
       # copy back labels to grouped data frame
       tmp <- sjlabelled::copy_labels(grps$data[[i]], dd)
 
-      dummy <- descr_helper(tmp, max.length)
+      dummy <- descr_helper(tmp, max.length)[, show]
       attr(dummy, "group") <- get_grouped_title(x, grps, i, sep = ", ", long = FALSE)
 
       # save data frame for return value
@@ -102,6 +127,7 @@ descr <- function(x, ..., max.length = NULL, weights = NULL, out = c("txt", "vie
 
   } else {
     dataframes <- descr_helper(dd, max.length)
+    dataframes <- dataframes[, intersect(colnames(dataframes), show)]
 
     # add class-attr for print-method()
     if (out == "txt")
@@ -112,6 +138,8 @@ descr <- function(x, ..., max.length = NULL, weights = NULL, out = c("txt", "vie
 
   # save how to print output
   attr(dataframes, "print") <- out
+  attr(dataframes, "encoding") <- encoding
+  attr(dataframes, "file") <- file
 
   dataframes
 }