+++ R/merge_df.R
@@ -9,7 +9,7 @@
 #'
 #' @return A full joined data frame.
 #'
-#' @details This function works like \code{\link[dplyr]{bind_rows}}, but preserves
+#' @details This function works like \code{\link[dplyr:bind]{dplyr::bind_rows()}}, but preserves
 #'   variable and value label attributes. \code{add_rows()} row-binds all data
 #'   frames in \code{...}, even if these have different numbers of columns.
 #'   Non-matching columns will be column-bound and filled with \code{NA}-values
@@ -57,9 +57,16 @@ add_rows <- function(..., id = NULL) {
     warning(sprintf("Value of `id` already exists as column name. ID column was renamed to `%s`.", id), call. = F)
   }
 
+  # remove variables with duplicated names
+
+  dat <- lapply(list(...), function(d) {
+    d[, unique(names(d)), drop = FALSE]
+  })
+
+
   # bind all data frames
 
-  x <- dplyr::bind_rows(..., .id = id)
+  x <- dplyr::bind_rows(dat, .id = id)
 
 
   # get attributes from all variables of original data frame