+++ R/word_wrap.R
@@ -5,12 +5,13 @@
 #'                labels / titles for plots or tables.
 #'
 #' @param labels Label(s) as character string, where a line break should be
-#'          inserted. Several strings may be passed as vector  (see 'Examples').
-#' @param wrap Maximum amount of chars per line (i.e. line length). If \code{wrap = Inf},
-#'          no word wrap will be performed (i.e. \code{labels} will be returned as is).
+#'   inserted. Several strings may be passed as vector  (see 'Examples').
+#' @param wrap Maximum amount of chars per line (i.e. line length). If
+#'   \code{wrap = Inf} or \code{wrap = 0}, no word wrap will be performed
+#'   (i.e. \code{labels} will be returned as is).
 #' @param linesep By default, this argument is \code{NULL} and a regular new line
-#'          string (\code{"\\n"}) is used. For HTML-purposes, for instance, \code{linesep}
-#'          could be \code{"<br>"}.
+#'   string (\code{"\\n"}) is used. For HTML-purposes, for instance, \code{linesep}
+#'   could be \code{"<br>"}.
 #' @return New label(s) with line breaks inserted at every \code{wrap}'s position.
 #'
 #' @examples
@@ -21,12 +22,17 @@
 #' @importFrom stats na.omit
 #' @export
 word_wrap <- function(labels, wrap, linesep = NULL) {
+  # infinite wrap? then return labels
+  if (is.infinite(wrap) | wrap == 0) return(labels)
+  # expressions can't be wrapped
+  if (is.expression(labels)) {
+    warning("Word wrap is not available for expressions.")
+    return(labels)
+  }
   # check if labels have NA values and remove them
   if (anyNA(labels)) labels <- as.character(stats::na.omit(labels))
   # check for valid value
   if (is.null(labels) || length(labels) == 0) return(NULL)
-  # infinite wrap? then return labels
-  if (is.infinite(wrap)) return(labels)
   # coerce to character, if factor
   if (!is.character(labels)) labels <- as.character(labels)
   # default line separator is \n